+++ wp_api_integration_tests/tests/test_block_patterns_err.rs
@@ -1,0 +1,11 @@
+use wp_api_integration_tests::prelude::*;
+
+#[tokio::test]
+#[parallel]
+async fn list_err_cannot_view_as_subscriber() {
+    api_client_as_subscriber()
+        .block_patterns()
+        .list_with_edit_context()
+        .await
+        .assert_wp_error(WpErrorCode::CannotView)
+}