+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [Block Patterns](https://developer.wordpress.org/rest-api/reference/block-patterns/) endpoint
 - [Block Types](https://developer.wordpress.org/rest-api/reference/block-types/) endpoint
 
 ## [0.2.0] - 2026-05-05